requirements.txt
# Core ETL
yfinance>=0.2.40
pandas>=2.2
numpy>=1.26,<3
SQLAlchemy>=2.0
PyMySQL>=1.1
python-dotenv>=1.0.1
tenacity>=8.2
pydantic>=2.6
tzdata>=2024.1; platform_system == "Windows"
python-json-logger>=2.0.7
tqdm>=4.66
